security tweaks.
Fixes the ID on the armory bulkheads so that they'll actually open/close with the rest.

Fixes the camera in the secure armory(marshal) to actually be on the sec network instead of the wall network

fixes the cameras on the wall to actually be on the wall network

renames the perimeter wall area from "pad" to "security main gate"

adds a couple more lockdown shutters to the security lobby(mostly to cover the door as well as the windows.

Ranger once again gets a choice in what armor they get instead of just getting an overcoat box.

Re-adds the 'colony lockdown' shutters that were missing from seemingly random windows/doors on the outside of the colony.

Adds a reinforced table/windoor to the prospector warehouse south wall opening so they can - yknow. Not have a hole in their warehouse when the bulkheads are down.

Slightly increases the resistance of blastdoors(16-20)