removes the mouth check from synths/fbps so they can perform eating(it does nothing but hey. flavor.)
A little bit of flavor, as a treat. Synths of all types can now eat/drink, unsurprisingly given that they aren't capable of processing reagents this results in their stomach filling and remaining at whatever level u have gluttonously filled it to. This is mostly just for flavor especially for synths/fbps that are supposed tohave mouths