re-enables supply pod drop
Once in a while a supply pod will fall from the sky, landing somewhere in the deep forest. It'll contain various loots and is likely(though not guaranteed) to carry some guardians- maybe these guardians are just some pests that hitched a ride, or perhaps they'll be a bit more dangerous.