Eris plasma port and balance (#4859)
* bwaaamp

* ''balance''

* last touches maybe

* lets not make these pointless

* lot cleaner way to comment stuff out huh

* lets not make this a church brand sunrise

* some small tweaks.

Gives extra recoil to the Purger was abit to easy to get it down to almost nothing.

Dominion made slightly less cell efficient.

* Update concilium.dm

Makes the concilium a genuine las LMG now taking power cells. Capacity similar to a normal LMG when using mid-tier large cells. increases plasteel cost to 40 (20 in a bioprinter), wield delay, slowdown and adds a unmaker block to compensate.

* Update psi_weptools.dm

woops suppose to be the same.

* sharing sound effect.

No balance changes. Distributes the new plasma sounds to the rest of the plasma weapons in the game between SI and church.

* Update construction.dm

woooops! Updates the Construction litany to the new gun materal costs prevent magicly producing mats