Removing Possum's Shit Code (#4911)
* Changing Possum's Shit Code

This will be changing the Kitsuhana cigarette's that were made as a very low blow towards Virgo's code and server while using the word 'retard' in the item description and making these cigarette's cause brain damage. Truly uncalled for behavior of Possum.

* Removing Possum's Shit Take

This changes the way that the Kitsuhana cigs interact with the player, removing the brain damage. Nerfing the serotrotium and mindbreaker to make it fit with the new theme of 'astrovan life'.

* Update fancy.dm

Fixing a typo