silicon language changes. (#4924)
Removes Opifexee and Native Plant from Silicons. As for why?

Native Plants desc suggests native plant is totally unknown in how it works. It's also a hivemind language and thus non-verbal.

As for bird talk. The Opifex are the most secretive racial society we have I simply can't see them sharing it in enough detail for silicons or anyone for that matter to be able to understand it. Ideally I'd prevent holopads from being able to but just not something I know how to do currently.