more event stuff. (#4916)
* suit

Adds a new RIG for event use.

* uoo

new spawner for events.

* Fixes a few bugs. Also buffs explo thing.

Slightly strengthens the explosion of the dead mans switch, so that it actually gibs you.

* More event stuff.

new ghostrole spawners and some various other stuff to support them. Event related

* Update ghostroles.dm

* Update icon.dmi

* Update mob.dmi

* new title screen

Art courtesy of Bamhalazam, depicts quote "the garden district, with the lift in the back and the church district nearby. "

* Update lobbyscreen.dm