telesci fixes/buffs
I always felt telesci was way to much of a pain. After some digging I noticed it's not actually been coded to account for are map.

Changes are as follows
Comments out telescis proc to check for SealedLevel. Since all of are z-levels are sealed this was maxing out the difficulty of telesci regardless of what level/area you target.

I've also midly bumped the pathfinding speed buff from cognition. a 4 to a 6. Might need to be even higher or not touched at all but I'd like to see this in practice.