senseable guest passes
comments out a few lines of code making guest pass computers default to deselected rather then all selected. 90% of the time this is used for letting outsiders in meaning no access! On the off time its not better to opt in to access then forget somewhere important opting out of access . Gone are the days of commanders giving a outsider access to the blackshield armory! (as funny as it was)