telesci fixes/buffs (#4940)
* telesci fixes/buffs

I always felt telesci was way to much of a pain. After some digging I noticed it's not actually been coded to account for are map.

Changes are as follows
Comments out telescis proc to check for SealedLevel. Since all of are z-levels are sealed this was maxing out the difficulty of telesci regardless of what level/area you target.

I've also midly bumped the pathfinding speed buff from cognition. a 4 to a 6. Might need to be even higher or not touched at all but I'd like to see this in practice.

* Update code/modules/telesci/telesci_computer.dm

Co-authored-by: CDB <87905328+cdb-is-not-good@users.noreply.github.com>

* Update toxins.dm

tosses in a super smol bugfix for gewaltine. It was giving CE_toxin 4 when its suppose to be a anti-toxin for the bugs.

* Update telesci_computer.dm

funny but over likely a bit to much. Reverts the buffs from cog to their old numbers. Mostly after seeing that lower difficulty teleports are pretty quick as is.

* adds the mini inhibitor

Adds a new machine.  developed by SI research essentially acts as a way to put a tele inhibitor in areas of high importance. Unlike full size ones does not contribute to global entropy protection. I'll get around to distributing these to the relevant areas once the map files are free from test merged PRs.

---------

Co-authored-by: CDB <87905328+cdb-is-not-good@users.noreply.github.com>