plasmachanges
A few changes to plasma. First off, it no longer causes toxin wounds EVERY tick and instead has a 6/100 chance to cause a wound as long as it's in your system per tick instead it merely offers high toxicity that grows with "dose"(a value that starts at 0 and rises to a max of the amount you've had added to your mob), this should prevent instances where the player touches stuff that has plasma in and INSTANTLY gets owned. Additiionally, contaminated stuff *also* no longer ticks plasma-inject every single tick, only doing so 1/10 times. Additionally, instead of causing toxin wounds on veins and only veins it now targets liver/kidneys, rolling a 3% probability check for liver and if that fails rolling another 3% check for kidneys. As a final note, the 3% prob check on the wounds is modified by a quarter of your dose(so if you have 4u of plasma it's an even 3%, whereas 1u of plasma is 0.75%.