new shield gun.
A very quick 'n sloppy thing for the shield who now have the option of a tweaked lascore with 20% better shot-efficiency and 30% worse firerate. Corpsmen/Sarge can take it as a primary or you can get it from the teleporter for a bit point cost.. A bit half assed but I've been meaning to add this for a while to help sure up the shields selection of energy weapons.