another new energy gun for BSC/WO.
Adds 'Gleam', a high powered high end laser carbine built off the laser-core platform. While losing a bit of portability, foldable stock and a bit of damage the Gleam provides even further improved efficiency over the blackshield lasercore(80 vs 100) and comes with a two round burst as well as a reflex scope and +2 dam mod when scoped(somewhat akin to the Ostwind.)

New sprite comes courtesy of Aurora who created the original 'blaster' sprite and which i shamelessly kitbashed to suit these needs.

Also removes the ammo pouches in the WO/COs lockers, they now get a pouch determined by their primary weapon kit(A tubular for the laser, a regular for the ballistics.)

Also also moves the COs gunbox to the...proper file, where it belongs with the rest of the gunboxes