FLATLINE SICKNESS 2.0 ELECTRIC BOOGALOO
Attempts to resolve an issue where one couldn't be "properly" revived when their brain dies. Nanopaste is once again able to rejuvenate a brain that has fully died. When repaired in this way, the individual permanently gains "Flatline Syndrome" a wasting weakness that serves as a particularly severe form of resurrection sickness that utterly nuked physical stats and applies severe added damage. This is offered as an alternative to my prior attempt to resolve people getting revived with destroyed brains(By making it impossible to do so). This allows you to get revived, at a severe cost. Obviously, you should strive to avoid having this done to you.

Also removes reference to filtering toxins out using nanopaste since that isn't really relevant to current med. Toxin buildup is repaired in other ways