new wanted notice. (#5017)
* FLATLINE SICKNESS 2.0 ELECTRIC BOOGALOO

Attempts to resolve an issue where one couldn't be "properly" revived when their brain dies. Nanopaste is once again able to rejuvenate a brain that has fully died. When repaired in this way, the individual permanently gains "Flatline Syndrome" a wasting weakness that serves as a particularly severe form of resurrection sickness that utterly nuked physical stats and applies severe added damage. This is offered as an alternative to my prior attempt to resolve people getting revived with destroyed brains(By making it impossible to do so). This allows you to get revived, at a severe cost. Obviously, you should strive to avoid having this done to you.

Also removes reference to filtering toxins out using nanopaste since that isn't really relevant to current med. Toxin buildup is repaired in other ways

* FLATLINE SICKNESS 2.0 ELECTRIC BOOGALOO

Attempts to resolve an issue where one couldn't be "properly" revived when their brain dies. Nanopaste is once again able to rejuvenate a brain that has fully died. When repaired in this way, the individual permanently gains "Flatline Syndrome" a wasting weakness that serves as a particularly severe form of resurrection sickness that utterly nuked physical stats and applies severe added damage. This is offered as an alternative to my prior attempt to resolve people getting revived with destroyed brains(By making it impossible to do so). This allows you to get revived, at a severe cost. Obviously, you should strive to avoid having this done to you.

Also removes reference to filtering toxins out using nanopaste since that isn't really relevant to current med. Toxin buildup is repaired in other ways

* Revert "FLATLINE SICKNESS 2.0 ELECTRIC BOOGALOO"

This reverts commit 9f54adb4893084648bce54cd002744047b1cb9b2.

* Revert "FLATLINE SICKNESS 2.0 ELECTRIC BOOGALOO"

This reverts commit 182125e3ff7be2d09b8e172ae14724e9f1a4dd6b.

* new wanted notice.

* Update stationobjs.dmi

* Update noticeboard.dm