buffs Entreaty
Makes the common cruciform litany Entreaty give exact coordinates as well as area.  I simply believe the most common places this is useful have far to big a area to not provide coordinates. Junk fields, maint, dungeons etc.  This is pretty difficult to test on a private server but it mimics the way death alarms handle it so SHOULD work.