new toys for the shield. (#5005)
* adds unnecessary mess to medical and also fixes welding for good this time.

fixes wselding protection to not be broke. Oops!

relocates the cold-loop for soteria cryo to a storage closet south of the actual tubes, this is a bit of a stop-gap bt should solve issues where cryo tubes are being cut off from their supply of the necessary cold air.  - Also changed the windows to polarizing windows in the cryo bay so you don't like...have random scientists coming up the elevator and seeing all the SICK AND DYING people.

Also renames a few areas to make them a bit clearer.

Lower Cryogenic storage - Lower Colony Elevator Lobby
Upper Cryogenic storage - Colony Cryosleep Storage
Cryogenics - Soteria Cryogenics Bay

This is hoped mostly to clear up confusion between the three. We may want in the future to add some distinction between Cryogenic(storage) and cryogenic(healing) but that's a lore issue.

* new shield gun.

A very quick 'n sloppy thing for the shield who now have the option of a tweaked lascore with 20% better shot-efficiency and 30% worse firerate. Corpsmen/Sarge can take it as a primary or you can get it from the teleporter for a bit point cost.. A bit half assed but I've been meaning to add this for a while to help sure up the shields selection of energy weapons.

* Revert "adds unnecessary mess to medical and also fixes welding for good this time."

This reverts commit 36a3bd09e03981fd310cad9c888d4755c9f7d276.

* everyone gets a lascore

* another new energy gun for BSC/WO.

Adds 'Gleam', a high powered high end laser carbine built off the laser-core platform. While losing a bit of portability, foldable stock and a bit of damage the Gleam provides even further improved efficiency over the blackshield lasercore(80 vs 100) and comes with a two round burst as well as a reflex scope and +2 dam mod when scoped(somewhat akin to the Ostwind.)

New sprite comes courtesy of Aurora who created the original 'blaster' sprite and which i shamelessly kitbashed to suit these needs.

Also removes the ammo pouches in the WO/COs lockers, they now get a pouch determined by their primary weapon kit(A tubular for the laser, a regular for the ballistics.)

Also also moves the COs gunbox to the...proper file, where it belongs with the rest of the gunboxes

* makes the onsuit sprite not always flash. oops.

* Update code/modules/projectiles/guns/energy/laser/lasercore.dm

Co-authored-by: benj8560 <benj856159@Gmail.com>

* few shield loadout tweaks.

No more HP ammo for CO, ammo is ball like every other BS kit. Primary kits for CO come with five mags like 90% of other shield kits. the 10% of shield primary kits that didnt' have 5 mags also have received them

---------

Co-authored-by: benj8560 <benj856159@Gmail.com>