CO armor tweaks.
CO armor has received some slight tweaking. Giving up 10 melee armor for ten energy armor to keep it more in line with being a military grade vest similar to the WOs armor(Which itself has 50/50/50 compared to the COs formerly 45/50/30 and now 35/50/40 armor.)

CO given a new helmet to match the above values because I get really annoyed by non matching pieces. Similar to WO armor it has a lil antenna on the side and some slight changes in the IFF markings to indicate it as a CO helmet- slightly better light and still retains the

Also fixes the alignment on the hooded helmets so they...actually line up with the eyes light-overlay.