MAYBE fixes issues with pain being insane
My testing seems to indicate that the maths that are meant to return only the organ efficiency for 1 limb are actually returning a far larger value resulting in the crazy ass numbers that we see when say....Aulvae/Exalts take pain damage. The maths now should account for the fact that it's TOTAL nerve efficiency, not just one limb. Obviously? Adding even one more nerve will cause significant extra pain.