Xenobotany QOL
Remapped the xenoflora lab for better workflow, hid the atmos secure room behind walls so that farm bots won't get stuck along with spreading out the room. Removed the animal enclosure because no one used it. Ever.  Generally just made the room nicer to look at and work in.

Also added an advanced grinder into the lab in place of the old atmos room replacing the industrial grinder that was in there, as the advanced one makes it far easier to keep the room organized and bottleless.