Fixes, more more fixes.
Rats are no longer immune to bullets.

Simplemobs are no longer immune to bullets.

Wield Delay Factor raised from 0 to 2, this will result in a slightly higher amount of time to wield(and thus higher vig needed to instant wield) but my testing didn't....really show that it was super significant. you can still instant wield a nagant with 25 vig(starting for BS) and i suspect this may need to be tweaked later.