Correction of: addition of an armor set with 2 color variants with special craftable variant and a new RIG suit for suit up
(correction of values to PROPER values and not old values added) Addition of a new mid tier armor suit for cargo's Sol Fed station since all armors are basic vests from starter stations or the assault suit from the kriosians. The new "sol fed" armor seeks to bridge the middle area between simple vest and end tier assault armor, simply balanced by putting it in a far (arguably 1 hour unlock) station while being affordable to the standard colonist and also adding a special variant which sacrifices modification of armor values with the ability to attach the chest rig craftable to give the armor a 5 cell internal inventory. The RIG suit seeks to add a new (slightly more expensive) light suit which offers high mobility and decent protection compared to heavier combat RIGs (low speed loss and slightly better armor than standard EVA RIG, similar to light RIG but atmos friendly). This adds new mob and icon sprites to accommodate all variants and pieces of the armors and RIGs, the respective changes to the armor code pages, and the changes to illegalstation1 and suitup station code pages to add both sets of armor to their intended stations. (note that the solfed armor price may need to be reviewed as a proper price point was hard to determine).