some combat PR balance tweaks
Buffalo has 1.5 AP mod up from 0.8

specialized RIG modules are generally rarer and more expensive, the standard level of ablative armor has gone from 10 to 5(that is to say, they now provide an extra 5 destroyable armor instead of 10, as this was maybe a bit much. combat suits retain their high ablation value.)

Industrial RIG has retained high ablation and gained a bit of max ablative, given that it's slow and kinda rare. makes it good for miners, but maybe not ideal for combat.

SWAT may no longer be disassembled.

if your armor divisor somehow becomes negative, instead of returning a 1 it return a 0.001, meanuing it is no longer a good idea to accept a negative armor divisor as it will unironically be unable to penetrate even jackboot tier armor.