QOL Fixes for Psionic Armor (#32)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
<details>
<summary>
	About The Pull Request
</summary>
<hr>

Adds two functionalities to psionic armor:
1. Clothing that used to be replaced entirely by the power is now
dropped on the floor when using the power.
2. Equipment in the suit storage slot is now reequipped on resolution of
the power.

This change also fixes some of the grammar and spelling in the power
descriptions and updates the messages displayed in chat to reflect the
fact that the psionic moth god no longer eats the psion's clothing when
using this armor power.
	
<hr>
</details>

## Changelog
:cl:
tweak: Psionic armor powers no longer eat your clothing and reequip
anything you had in your suit slot on use.
:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Co-authored-by: valkyria-gk <177847920+valkyria-gk@users.noreply.github.com>