Vote to power the colony, take two. (#46)
## About The Pull Request
Adds a simple vote whether or not to power the station. Useful for
lowpop rounds with no engineers to keep the lights on. Basically just
ported forward from Occulus minus the shields, since the colony doesn't
have any. On passing, the vote triggers the same proc as an admin
hitting the 'power all SMES' thing.

Useful for when we want to perform Shakespeare in the Park instead of
Shakespeare in the Dark.

## Changelog
:cl:
code: new vote type to fully power the colony
/:cl: