Altyn helmet fix (#53)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

A very simple fix that only occurs for the Altyn helmet. Generally,
there are 4 types of helmet (with variations) that has an option to
"flip up" the face mask.

For all of them, the armor value consistently go down when the facemask
is up, and go up when the face mask is down to protect you _EXCEPT_ the
Altyn and it's variations such as Maska.
* This fixes that problem, since `armor_up` is the alternative value
called when the face plate is pushed up and out of the way to show the
person their face.
<hr>

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
	
## Changelog
:cl:
fix: fixes the altyn helmet armor value
/:cl: