Added ability to Direct Recycle (#50)
I-Lathe and Bioprinter have been given the ability to direct recycle via
a right-click verb. When enabled items that can be recycled will do so
instantly upon clicking the lathe. This can be turned off as well to
prevent accidentally recycling objects in hand.

## About The Pull Request

I-Lathe and Bioprinter have been given the ability to direct recycle via
a right-click verb. When enabled items that can be recycled will do so
instantly upon clicking the lathe. This can be turned off as well to
prevent accidentally recycling objects in hand.

Also a sprite given by Emilitia Ennehrt in order to denote when it is or
is not active.
	About The Pull Request


## Changelog
:cl:
tweak: tweaked the functionality of recycling via the I-Lathe and
Bioprinter
imageadd: added a button sprite for the I-Lathe and Bioprinter
/:cl:

---------

Co-authored-by: valkyria-gk <177847920+valkyria-gk@users.noreply.github.com>