Gun Sprite Fix - Part One (#71)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About the Pull Request

Some gun sprites are broken, straight up. They've implemented some
systems to change gun sprites according to the state of the gun such as
level of cell charge or if the gun is empty or not (blinking light).
This has caused some sprites to break since not all of them were
properly updated to use the new system name call.

This is part one of an attempt to fix all sprites from not working to
working properly.

List of Changes in Part One
<details>
  <summary>
    Expand for list
  </summary>

* Antebellum - previously no wield sprite
* SST Abengate - previously no held of wield sprite
* Excubitor - previously no sprite on back or suit slot
* Firestorm - previously no sprite for charge less than 100% and more
than 0%
* SST Format Bound - previously no held or wield sprite
* Temperature Gun - previously no wield sprite and sprite on back or
suit slot
* Unknown Plasma Gun (from Strangers) - previously no wield sprite
* Makeshift laser carbine - previously no wield sprite
* Martyr - previously no held or wield sprite
* Serenity - previously no held or wield sprite
* Mana from Heaven - previously no held or wield sprite
* Nuclear Particle Rifle - previously no sprite on back or suit slot
* Dissolver - previously no sprite on back or suit slot
* Pinion - previously no held of wield sprite
</details>

## Changelog
:cl:
fix: fixes some gun sprites
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->