Solarian armor added helmet and mask for full set no kit bashing
addition of a helmet and gas mask for the chemical corps Solarian armor so there is a full set of armor that matches or looks better than mixing and matching different helmets to fit the vibe of the armor. Mask is a general gasmask but with 100 rad protection, helmet is essentially the ballistic mask from sec disk minus one point and some added rad and bio protection to fit the armors intention. Both mask and helmet are added to arau favored inventory next to the Solarian armor