Account Database Tweak
Allows Premiers/BST/Centcom access cards to see account pins in the Account Database machine.

It also prints the account pin on the paper.

I figure if you're accessing accounts at that level you're trusted with the Pin to begin with, and the printed papers should be treated as highly confidential anyways.

I'm scared of NanoUI.