One line fix to make LOOC stop popping up the chat bubble (#120)
## About The Pull Request
Currently, in Hotkey mode, when you press O to type in OOC, there is no
typing indicator popping up on your mob. However, when you press L to
type in LOOC, a typing indicator pops up.

This PR simply makes it so that LOOC will not cause a typing indicator
to appear.

## Changelog
:cl:
fix: LOOC should no longer cause a typing indicator to appear on your
character.
/:cl: