Change rules to a link system (#132)
## About The Pull Request
Change the default link in example config to our codebase. Add rulesurl
as a variable so that we can just link to a much more neatly formatted
link

## Changelog
:cl:
server: Rules is no longer stored in rules.html and is now stored in
rulesurl as a link.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->