Upstream Update #1 - Add EAR (Psionic Catalyst Machine) & Litanies (#128)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This PR cherry picks a bunch of stuff from upstream including the new
psionic catalyst transmutation thing (the EAR) and a bunch of new
litanies for churchies.


![dreamseeker_DZrRxjCNff](https://github.com/user-attachments/assets/80867c72-8817-4d68-a915-0c6563fb2ed9)

![dreamseeker_zzQ1IGPtiC](https://github.com/user-attachments/assets/99c6900b-1941-42b8-948b-f641a985f65c)


## Changelog
:cl:
add: Psionic catalyst transmutation via the EAR in research underground,
below the medical room (The medical room has been shrunk to make room
for it)
add: Bunch of new litanies to play with
/:cl: