Crafting step is no longer globally shared  (#131)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Closes #122. This is an issue that is also in Eris code because of how
the building variable works.

The building variable is on the **global** crafting step, meaning that
anyone doing one step will block **everyone** globally from doing that
step.

This changes it by removing the variable and making it so that apply
simply change the target - if it exists, in question's step to the
target step. It also adds a safety check when target exists to avoid
redundantly applying the same step.

As a side effect, this means mass crafting is also possible by the same
person on the same recipe. If this is abused to a point that it affects
server performance or leave junk items - I'll consider adding in a per
mob maximum check of do after or something like that.

## Changelog
:cl:
fix: Crafting one step on the server no longer blocks the entire step on
the entire server.
tweak: Above also means the same person can craft the same item multiple
time. Do not abuse this.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->