Show extra inventory and recommendation favor threshold separately. (#127)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Tweak the UI of the trading program to display the extra inventory
(Hidden Inventory) and the recommendation threshold separately for
better transparency.

Delete unused trade_catalog.tmpl

Originally I wanted to show the amount of recommendations needed for
trade stations, but after dealing with NanoUI for 2+ hours I've decided
I'd rather butcher the whole thing and convert it into TGUI than deal
with it again. But since I don't have infinite time I am leaving that
for later.

Remove "Favor" from the Merchant screen as I think it should be known
and it was kinda hard to work with NanoUI.

New UX:

![c8V6UA0O13](https://github.com/user-attachments/assets/1c026b6f-b96d-425b-a3ff-6c76fe2ebdc8)

![pc4LnDqglR](https://github.com/user-attachments/assets/57f4643d-50ad-459e-a37b-8732785ea254)

## Changelog
:cl:
tweak: Trading Program now shows the amount of favor needed to unlock
the hidden inventory and recommendations separately, instead of just the
highest of the two to avoid confusing players.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->