Wendigo Rename (#154)
Renames Wendigo to Timberstag.

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
<details>
<summary>
Renames Wendigo to Timberstag. It's an easy fix for cultural
sensitivity, plus I think Timberstag is a cooler name anyway.
</summary>
<hr>

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
	
<hr>
</details>

## Changelog
:cl:
tweak: Renamed a tengolo evolution to Timberstag.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->