New seaslug-theme ears and tails (#155)
## About The Pull Request

![Y8qV4BgADP](https://github.com/user-attachments/assets/917c528d-4192-4412-881a-86e904fa3275)

![IxEZYRHGOI](https://github.com/user-attachments/assets/7b722567-3ba3-4f34-aa1a-a1742f4ef8b1)


## Changelog
:cl:
add: New Moyo antennae and tails are available in the character creator.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->