Fix map autodoc and add beep noise on scan (#162)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Fix map autodoc and add beep noise on scan. 

The map autodoc, when created, doesn't link to the medical account
because it is not there yet. Add a beep sound on scan so that user knows
when it has scanned and have found no fixable injuries.

## Changelog
:cl:
fix: Actually fix the map autodoc by autolinking to medical account
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->