Update contributing.md, readme and code of conduct (#163)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This PR update our description a little bit, add a Code of Conduct and
refine the existing Contributing.md (I hope it works) with some new code
standards and stuffs, such as not adding NanoUI and not adding so-called
"Modular Code". Also move their folders.

## Changelog
:cl:
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->