BirdsOfWar01
Let slip the birds of war

Add 3 new unique mobs
Ciriklo Gunner - This Ciriklo is strapped into a Wheelchair with a Bastard SMG being pushed by his support gunner Ciriklo, making this unit a two for one deal.  When the gunner fires, there is a 10% chance the support gunner Ciriklo flings off the back of the wheelchair.  Ciriklo Gunners without Support Gunners get a drastic reduce to their speed, effectively turning them into living turrets.

Ciriklo Cannon - The same as above, except instead of a Bastard SMG, they use a makeshift China Lake that shoots Frag shells with a 40% chance of knockback.

Ciriklo Rider - This Ciriklo rides the back of a mighty Hell Pig allied with the Ciriklo by default.  It has the same stats as a Hell Pig, making this the tankiest unit Ciriklo now have.  Once killed, the rider on it's back will hop off and join the fight!

A lot of testing, balance, and visual tweaking has gotten this functional and I'm ready to present these new units for the on-going Ciriklo arc

THIS UNITS ARE NOT TO BE USED UNTIL WILSON/HEADMIN ALLOWS IT